+++ packages/impulse-ui-core/example/vite.config.js
@@ -1,7 +1,0 @@
-import ReactPlugin from '@vitejs/plugin-react';
-import { defineConfig } from 'vite';
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [ReactPlugin({})],
-});