+++ packages/impulse-ui-sidebar/src/components/sidebar/sidebar-item/theme.map.ts
@@ -1,0 +1,9 @@
+import { ComponentMap } from '@impulse-ui/types';
+
+const sidebarItemComponentMap: ComponentMap[] = [
+  { key: 'buttonStyle' },
+  { key: 'iconStyle' },
+  { key: 'typographyStyle' },
+];
+
+export { sidebarItemComponentMap };