+++ packages/impulse-ui-sidebar/src/components/sidebar/sidebard-section-divider/index.ts
@@ -1,0 +1,2 @@
+export * from './SidebarSectionDivider';
+export * from './theme.map';