+++ packages/impulse-ui-sidebar/src/components/sidebar/theme.map.ts
@@ -1,0 +1,5 @@
+import { ComponentMap } from '@impulse-ui/types';
+
+const sidebarComponentMap: ComponentMap[] = [{ key: 'containerStyle' }];
+
+export { sidebarComponentMap };