+++ packages/impulse-ui-toolkit/CHANGELOG.md
@@ -5,7 +5,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [2.1.0-beta.0](https://github.com/apolyanov/impulse-ui/compare/v2.0.7-beta.0...v2.1.0-beta.0) (2023-10-10)
 
-
 ### Features
 
-* Adding Impulse UI Toolkit package ([c63bdd7](https://github.com/apolyanov/impulse-ui/commit/c63bdd7dbfdbdd74e633cd6a49783036e8d69d3f))
+- Adding Impulse UI Toolkit package ([c63bdd7](https://github.com/apolyanov/impulse-ui/commit/c63bdd7dbfdbdd74e633cd6a49783036e8d69d3f))